# Conflicts:
#	src/main/java/com/ark/demo/controllers/RequestController.java
#	src/main/java/com/ark/demo/controllers/ResponseController.java
#	src/main/java/com/ark/demo/controllers/ThreadController.java
#	src/main/java/com/ark/demo/models/Request.java
#	src/main/java/com/ark/demo/models/Response.java
#	src/main/java/com/ark/demo/models/Thread.java
#	src/main/java/com/ark/demo/models/User.java
#	src/main/java/com/ark/demo/models/data/ThreadRepository.java
#	src/main/java/com/ark/demo/models/dto/CreateRequestFormDTO.java
#	src/main/resources/templates/fragments/requestFragments.html
#	src/main/resources/templates/requestTemplates/createRequest.html
#	src/main/resources/templates/requestTemplates/requestConfirmation.html
#	src/main/resources/templates/requestTemplates/userRequests.html
#	src/main/resources/templates/requestTemplates/viewRequest.html

+++ src/main/java/com/ark/demo/controllers/UserController.java
@@ -8,6 +8,7 @@ import com.ark.demo.models.dto.DeleteFormDTO;
 import com.ark.demo.models.dto.EditProfileFormDTO;
 import com.ark.demo.models.dto.UpdatePasswordFormDTO;
 import com.ark.demo.models.dto.ViewProfileDTO;
+import com.ark.demo.services.EmailService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -17,10 +18,13 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import javax.mail.MessagingException;
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
+import java.io.UnsupportedEncodingException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
+import java.util.Optional;
 
 import static java.util.Objects.isNull;
 
@@ -36,15 +40,19 @@ public class UserController {
     @Autowired
     private UserDetailsRepository userDetailsRepository;
 
+    @Autowired
+    private EmailService emailService;
+
     @GetMapping
     public String displayUserProfile(HttpServletRequest request, Model model){
         User user = authenticationController.getUserFromSession(request.getSession());
         if(isNull(user)){
-            return "redirect:../login";
+            return "redirect:/login";
         }
         ViewProfileDTO viewProfileDTO = new ViewProfileDTO();
         viewProfileDTO.setUserDetails(user.getUserDetails());
         viewProfileDTO.setDateCreated(formatDateAsString(user.getDateCreated()));
+        model.addAttribute(user);
         model.addAttribute("title","View Profile");
         model.addAttribute(viewProfileDTO);
         return "userTemplates/viewProfile";
@@ -64,7 +72,7 @@ public class UserController {
     }
 
     @PostMapping("/editProfile")
-    public String processUpdatePasswordForm(@ModelAttribute @Valid EditProfileFormDTO editProfileFormDTO, Errors errors, HttpServletRequest request, Model model){
+    public String processUpdatePasswordForm(@ModelAttribute @Valid EditProfileFormDTO editProfileFormDTO, Errors errors, HttpServletRequest request, Model model) throws MessagingException, UnsupportedEncodingException {
         if(errors.hasErrors()){
             model.addAttribute("title","Edit Profile");
             model.addAttribute(editProfileFormDTO);
@@ -84,6 +92,11 @@ public class UserController {
         userDetails.setState(editProfileFormDTO.getUserDetails().getState());
         userDetails.setZipcode(editProfileFormDTO.getUserDetails().getZipcode());
         userDetails.setPhoneNumber(editProfileFormDTO.getUserDetails().getPhoneNumber());
+        if(!userDetails.getEmailAddress().equals(editProfileFormDTO.getUserDetails().getEmailAddress())){
+            userDetails.setUid(editProfileFormDTO.getUserDetails().getEmailAddress());
+            userDetails.setEmailVerified(false);
+            emailService.sendRegistrationEmail(editProfileFormDTO.getUserDetails().getEmailAddress(), userDetails.getUid());
+        }
         userDetails.setEmailAddress(editProfileFormDTO.getUserDetails().getEmailAddress());
         user.setLocation(editProfileFormDTO.getLocation());
         userRepository.save(user);
@@ -143,6 +156,10 @@ public class UserController {
 
     @PostMapping("/deleteProfile")
     public String processDeleteProfileForm(@ModelAttribute @Valid DeleteFormDTO deleteFormDTO, HttpServletRequest request){
+        User user = authenticationController.getUserFromSession(request.getSession());
+        if(isNull(user)){
+            return "redirect:../login";
+        }
         if(deleteFormDTO.getConfirm().toLowerCase().equals("yes")){
             userDetailsRepository.delete(deleteFormDTO.getUserDetails());
             userRepository.delete(deleteFormDTO.getUser());
@@ -155,4 +172,5 @@ public class UserController {
         SimpleDateFormat simpleDateFormat = new SimpleDateFormat(pattern);
         return simpleDateFormat.format(date);
     }
+
 }
